(sew ((tie torn paper proto type suit per power
((((i feel it shows a tie and a thunder, a strike and a break, a cloth and a paper and a thought and a feel?

($?( i feel maybe i can show a person wearing a tie that breaksthrough and breaks some truth and maybe breakdances. ? (Power Icon by @LlubNek, A lightening bolt in negative space on a dark blue background in a rounded square. Created as an icon for a power adapter.

sew? wWw.pinterest.com/suitperpower/tie-torn-paper-proto-type-suit-per-power/ -/?/!/+ wWw.pinterest.com/seriousest/ssserious-profit-serious-biz-serious-style-srs-fas/

sew sew?

sew sew sew?